+++ server.py
@@ -19,6 +19,14 @@ import time
 
 import liblo
 
+import lighting
+import water
+import sound
+
+from lighting import Lighting
+from water import Water
+from sound import SoundEffects
+
 CONSOLE_LOG_LEVEL = logging.DEBUG
 FILE_LOG_LEVEL = logging.DEBUG
 LOG_FILE = 'amcpserver.log'
@@ -40,6 +48,10 @@ ch.setFormatter(formatter)
 logger.addHandler(fh)
 logger.addHandler(ch)
 
+# Deal with crazy logging
+lighting.logger = logger
+water.logger = logger
+sound.logger = logger
 
 class AMCPServer(liblo.ServerThread):
     def __init__(self, port):
@@ -95,107 +107,6 @@ class AMCPServer(liblo.ServerThread):
     # @liblo.make_method(None, None)
     # def fallback(self, path, args):
     #     print "received unknown message '%s' Args: %s" % (path, args)
-
-
-class Water():
-    """Controls rain, mist, etc"""
-
-    def rain(self, toggle):
-        logger.debug('action="rain", toggle="%s"' % toggle)
-        if toggle:
-            logger.info('action="rain", toggle="on"')
-        else:
-            logger.info('action="rain", toggle="off"')
-
-    def make_it_rain(self):
-        logger.info('action="make_it_rain"')
-        self.rain(True)
-        time.sleep(5000)
-        self.rain(False)
-        pass
-
-    def all_the_rain(self):
-        # PiGPIO.send(1,1)
-        logger.info('action="all_the_rain"')
-        pass
-
-    def rain_timer(self, length):
-        # PiGPIO.send(1, 1)
-        # sleep(length)
-        # PiGPIO.send(1,0)
-        pass
-
-
-class PiGPIO():
-    """Controls water (pumps and valves)"""
-
-    def send(self, pin_num, value):
-        """Send value (1/0) to pin_num"""
-        logger.debug('action="send_rpi_gpio", pin_number="%i", value="%i"'
-                     % (pin_num, value))
-        pass
-
-
-class Lighting():
-    """Should be able to translate something into appropriate index for OPC"""
-    def strobe(self):
-        # quick strobe
-        pass
-
-    def flood_lights(self, light_num, intensity):
-        # Turn on light_num at intensity
-        pass
-
-
-class OPC():
-    """Controls all the lights! LED, lightning, etc."""
-    def send(self, index, r, g, b):
-        """ Send r, g, b to index LED
-        """
-        pass
-
-
-class SoundEffects():
-    """Play different sound effects.
-
-    Probably want to index these sounds somehow? Config file?"""
-    def __init__(self):
-        self.so = SoundOut()
-
-    def press_play(self, sound_file):
-        self.so.play(sound_file)
-        logger.info('action="play_sound", soundfile="%s"' % (sound_file))
-
-    def thunder(self, press):
-        sound_file = os.path.join(MEDIA_DIRECTORY, 'thunder_hd.mp3')
-        if press:
-            self.press_play(sound_file)
-
-    def rain(self, press):
-        sound_file = os.path.join(MEDIA_DIRECTORY, 'rainsounds.wav')
-        if press:
-            self.press_play(sound_file)
-
-    def its_raining_men(self, press):
-        sound_file = os.path.join(MEDIA_DIRECTORY, 'its_raining_men.wav')
-        if press:
-            self.press_play(sound_file)
-
-
-class SoundOut():
-    """mplayer to RPi audio out"""
-    def __init__(self):
-        my_system = platform.system()
-        logger.debug('action=init_soundout, system="%s"' % my_system)
-        if my_system == 'Darwin':  # OS X
-            self.player = '/usr/bin/afplay'
-        elif my_system == 'Linux':
-            self.player = '/usr/local/bin/mplayer'  # is this always correct?
-
-    def play(self, soundfile):
-        # play that funky soundfile
-        subprocess.Popen([self.player, soundfile])
-
 
 if (__name__ == "__main__"):
     try:
